Fix dasdfmt during graphical kickstart. (#1269863)
This fixes a traceback that occurred if you attempted to run ks
with --cdl option specified; the way the list of LDL DASDs was
collected was faulty.

This also fixes an extraneous, but related issue; basically had to find a way to
update the devicetree with the device changes, otherwise the installer thinks
there is not enough space or worse. Decided the simplest way to do this was to
remove the device completely once it's been formatted. This gets rid of all
traces of the "old" (LDL or unformatted) device. Then after all formatting is
done, call devicetree.populate to detect the active (but missing) devices and
add them to the devicetree again, in their proper state.

Resolves: rhbz#1269863