Add device id to dasdfmt screen. (#1269174)
Since device id's do not always map to device names as expected
(e.g. dasda -> 0.0.0200, dasdb -> 0.0.0201, etc.), include the device
id on the dasdfmt screen so that it's clear which devices are going to
be formatted.

This required a couple of changes to the storage spoke as well since we
were only passing a list of device names to the dasdfmt dialog; now it
passes a list of DASDDevice objects, so the dasdfmt dialog can access
the device id attr.

See: https://sbueno.fedorapeople.org/1269174.png
(Although, only one space between devname and devid, not two.)

Resolves: rhbz#1269174