Combine formatting of LDL and unformatted DASDs. (#1316088)
Previously the dasdfmt screen was shown twice, once for all unformatted DASDs,
and once for all LDL DASDs. But that's annoying, and the screen text didn't
actually identify if it was unformatted or LDL DASDs that were being formatted
at the moment. So just combine it all into one list. Text-mode does this as
well, with the fix for 1259437 applied.

Also, ldldasds was using some flawed logic to gather that list (passing in the
entire list of DASDs instead of only those selected), so fixed that as well.

Resolves: rhbz#1316088