Provide a default install class.
Since we removed the base install class from available install
classes, anaconda can be unable to find an install class to use and
fail. This can happen for example in lorax with a product name
specified to a custom name. Because of that, we should provide a
default install class that will be used in that case.

Related: rhbz#1412159