Replace filterSupportedLangs and filterSupportedLocales
The methods for filtering supported languages and locales were removed
from the install classes, because the install classes are going to be
replaced with configuration files, so they cannot contain any methods.

The logic is moved to pyanaconda.localization and the filtering is
done via methods is_language_supported and is_locale_supported of the
payload object.