Create a new payload handler mixin class for tests
This mixin class will handle payload module creation, mocking
publishing of the handler etc.
In general all the functionality which have to be there for all handlers
of payload module.

This class was separated to another file without a _test.py suffix on
purpose. This is payload module specific, so it shouldn't be part of the
general __init__.py file and it can't be part of the
module_payload_test.py file because importing of that in the other test
files would mean launching imported tests in that file (again).